+++ src/components/Blog/BlogList/BlogListItem/BlogListItem.tsx
@@ -12,12 +12,47 @@ import BoxBorders from "../../../SpriteSheet/BrushSvgs/BoxBorders";
 import { P, Heading, HeadingTag } from "../../../Typography";
 import AspectRatio from "../../../AspectRatio";
 import OakImage from "../../../OakImage";
+import { ResolveOakHrefProps } from "../../../../common-lib/urls";
+
+type BlogListItemContentType = "blog-post" | "webinar";
+
+const getItemLinkProps = (props: BlogListItemProps): ResolveOakHrefProps => {
+  switch (props.contentType) {
+    case "blog-post":
+      return {
+        page: "blog",
+        slug: props.slug,
+      };
+    case "webinar":
+      return {
+        page: "webinars",
+        slug: props.slug,
+      };
+  }
+};
+const getItemCategoryLinkProps = (
+  props: BlogListItemProps
+): ResolveOakHrefProps => {
+  switch (props.contentType) {
+    case "blog-post":
+      return {
+        page: "blog-index",
+        category: props.category.slug,
+      };
+    case "webinar":
+      return {
+        page: "webinars-index",
+        category: props.category.slug,
+      };
+  }
+};
 
 export type BlogListItemProps = {
   titleTag: HeadingTag;
   title: string;
-  snippet: string;
-  href: string;
+  summary: string;
+  slug: string;
+  contentType: BlogListItemContentType;
   category: BlogWebinarCategory;
   date: string;
   withImage?: boolean;
@@ -29,7 +64,7 @@ export type BlogListItemProps = {
 );
 
 /**
- * Contains an image, title, and text snippet.
+ * Contains an image, title, and text summary.
  * The component contains a link styled as a button, which
  * whose click target stretches across the entire component.
  * The title tag (h1, h2, ...) is passed as a prop.
@@ -38,14 +73,13 @@ const BlogListItem: FC<BlogListItemProps> = (props) => {
   const {
     titleTag,
     title,
-    snippet,
-    href,
+    summary,
     category,
     date,
     withImage,
     mainImage,
-    contentType,
     thumbTime,
+    contentType,
   } = props;
 
   const {
@@ -117,8 +151,7 @@ const BlogListItem: FC<BlogListItemProps> = (props) => {
         >
           <OakLink
             {...categoryHoverProps}
-            page="blog-index"
-            category={category.slug}
+            {...getItemCategoryLinkProps(props)}
             focusStyles={["underline"]}
             $font="heading-7"
             $color="hyperlink"
@@ -132,8 +165,7 @@ const BlogListItem: FC<BlogListItemProps> = (props) => {
         <Heading tag={titleTag} $font={"heading-5"} $mt={8}>
           <OakLink
             {...primaryTargetProps}
-            page={null}
-            href={href}
+            {...getItemLinkProps(props)}
             htmlAnchorProps={{ title }}
             focusStyles={["underline"]}
             isHovered={cardIsHovered && !categoryIsHovered}
@@ -142,7 +174,7 @@ const BlogListItem: FC<BlogListItemProps> = (props) => {
           </OakLink>
         </Heading>
         <P $font={"body-3"} $mt={8} $mb={[8, 0]}>
-          <LineClamp lines={2}>{snippet}</LineClamp>
+          <LineClamp lines={2}>{summary}</LineClamp>
         </P>
       </Flex>
     </Flex>