+++ src/components/Button/CopyLinkButton/CopyLinkButton.test.tsx
@@ -5,11 +5,6 @@ import renderWithTheme from "../../../__tests__/__helpers__/renderWithTheme";
 
 import CopyLinkButton from "./CopyLinkButton";
 
-const useFeatureFlags = jest.fn(() => ({ enabled: {} }));
-jest.mock("posthog-js/react", () => ({
-  useFeatureFlags: () => useFeatureFlags(),
-}));
-
 describe("Copy link button", () => {
   it("renders", () => {
     const { getByRole } = renderWithTheme(
@@ -43,19 +38,5 @@ describe("Copy link button", () => {
 
     const clickedButton = getByLabelText("Copied to clipboard");
     expect(clickedButton).toBeInTheDocument();
-  });
-
-  it("renders IconButton if feature flag 'test-feature-flag' enabled", () => {
-    useFeatureFlags.mockImplementationOnce(() => ({
-      enabled: { "test-feature-flag": true },
-    }));
-
-    const { getByRole } = renderWithTheme(
-      <ToastProvider>
-        <CopyLinkButton />
-      </ToastProvider>
-    );
-
-    expect(getByRole("button")).toBeInTheDocument();
   });
 });